+++ examples/2_5_Triangular)graph.pdf
