+++ examples/6_3_graph.pdf
