+++ examples/6_6_graph.pdf
