+++ examples/classical_run_20250327/confusion_matrix.png
