+++ examples/classical_run_20250327/digit_collage.png
