+++ examples/classical_run_20250327/full_report.tex
