+++ examples/classical_run_20250327/performance_table.tex
