+++ examples/classical_run_20250327/training_curves.png
