+++ examples/classical_run_20250327/variance_explained.png
