+++ examples/hybrid_run_20250328/confusion_matrix.png
