+++ examples/hybrid_run_20250328/training_curves.png
